Renamed and built proper data classes
