--- app/src/main/java/com/tryden/planets/model/BasicDetails.kt
@@ -1,6 +1,0 @@
-package com.tryden.planets.model
-
-data class BasicDetails(
-    val mass: String,
-    val volume: String
-)
