--- app/src/main/java/com/tryden/planets/model/ImgSrc.kt
@@ -1,6 +1,0 @@
-package com.tryden.planets.model
-
-data class ImgSrc(
-    val img: String,
-    val imgDescription: String
-)
